wip: Api controller

+++ Controllers/AlertAsteroid.cs
@@ -5,8 +5,8 @@ namespace asteroidalert
     public class AlertAsteroid
     {
         public DateTime Fecha { get; set; }
-        public int Velocidad { get; set; }
-        public int Diametro { get; set; }
+        public decimal Velocidad { get; set; }
+        public double Diametro { get; set; }
         public string Nombre { get; set; }
         public string Planeta { get; set; }
     }
